+++ docs/src/index.tsx
@@ -3,8 +3,7 @@ import * as ReactDOM from 'react-dom'
 import { AppContainer } from 'react-hot-loader'
 
 // TODO make themes a monorepo of packages
-import { fontFaces, staticStyles, theme } from 'src/themes/teams'
-import { Provider } from '@stardust-ui/react'
+import { Provider, themes } from '@stardust-ui/react'
 
 import Router from './routes'
 
@@ -18,7 +17,7 @@ document.body.appendChild(mountNode)
 const render = NewApp =>
   ReactDOM.render(
     <AppContainer>
-      <Provider theme={theme} staticStyles={staticStyles} fontFaces={fontFaces}>
+      <Provider theme={themes.teams}>
         <NewApp />
       </Provider>
     </AppContainer>,
@@ -34,13 +33,13 @@ if (__DEV__) {
   if (module.hot) {
     module.hot.accept('./routes', () => {
       // restore scroll
-      const { scrollLeft, scrollTop } = document.scrollingElement
+      const { scrollLeft, scrollTop } = document.scrollingElement!
       ReactDOM.unmountComponentAtNode(mountNode)
 
       try {
         render(require('./routes').default)
-        document.scrollingElement.scrollTop = scrollTop
-        document.scrollingElement.scrollLeft = scrollLeft
+        document.scrollingElement!.scrollTop = scrollTop
+        document.scrollingElement!.scrollLeft = scrollLeft
       } catch (e) {
         console.error(e)
       }