+++ src/themes/teams/components/Button/buttonStyles.ts
@@ -4,7 +4,7 @@ import { disabledStyle, truncateStyle } from '../../../../styles/customCSS'
 
 const buttonStyles: IComponentPartStylesInput = {
   root: ({ props, variables }: { props: any; variables: any }): ICSSInJSStyle => {
-    const { circular, disabled, fluid, icon, iconPosition, type } = props
+    const { circular, disabled, fluid, type } = props
     const primary = type === 'primary'
     const secondary = type === 'secondary'
 
@@ -12,6 +12,7 @@ const buttonStyles: IComponentPartStylesInput = {
       height,
       minWidth,
       maxWidth,
+      color,
       backgroundColor,
       backgroundColorHover,
       circularRadius,
@@ -30,27 +31,43 @@ const buttonStyles: IComponentPartStylesInput = {
       height,
       minWidth,
       maxWidth,
+      color,
       backgroundColor,
-      display: 'inline-block',
+      display: 'inline-flex',
+      justifyContent: 'center',
+      alignItems: 'center',
       position: 'relative',
       padding: `0 ${pxToRem(paddingLeftRightValue)}`,
       margin: `0 ${pxToRem(8)} 0 0`,
       verticalAlign: 'middle',
       borderRadius: pxToRem(2),
-      borderWidth: 0,
 
-      ...truncateStyle,
+      borderWidth: `${secondary ? (circular ? 1 : 2) : 0}px`,
+      cursor: 'pointer',
+      ':hover': {
+        backgroundColor: backgroundColorHover,
+      },
 
-      ...(icon &&
-        (iconPosition
-          ? {
-              display: 'inline-flex',
-              justifyContent: 'center',
-            }
-          : {
-              minWidth: height,
-              padding: 0,
-            })),
+      ...(primary && {
+        color: typePrimaryColor,
+        backgroundColor: typePrimaryBackgroundColor,
+        borderColor: typePrimaryBorderColor,
+        ':hover': {
+          color: typePrimaryColor,
+          backgroundColor: typePrimaryBackgroundColorHover,
+        },
+      }),
+
+      ...(secondary && {
+        color: typeSecondaryColor,
+        backgroundColor: typeSecondaryBackgroundColor,
+        borderColor: typeSecondaryBorderColor,
+        ':hover': {
+          color: typeSecondaryColor,
+          borderColor: 'transparent',
+          backgroundColor: typeSecondaryBackgroundColorHover,
+        },
+      }),
 
       ...(circular && {
         minWidth: height,
@@ -63,36 +80,20 @@ const buttonStyles: IComponentPartStylesInput = {
         maxWidth: '100%',
       }),
 
-      ...(disabled
-        ? disabledStyle
-        : {
-            borderWidth: `${secondary ? (circular ? 1 : 2) : 0}px`,
-            cursor: 'pointer',
-            ':hover': {
-              backgroundColor: backgroundColorHover,
-            },
-
-            ...(primary && {
-              color: typePrimaryColor,
-              backgroundColor: typePrimaryBackgroundColor,
-              borderColor: typePrimaryBorderColor,
-              ':hover': {
-                backgroundColor: typePrimaryBackgroundColorHover,
-              },
-            }),
-
-            ...(secondary && {
-              color: typeSecondaryColor,
-              backgroundColor: typeSecondaryBackgroundColor,
-              borderColor: typeSecondaryBorderColor,
-              ':hover': {
-                borderColor: 'transparent',
-                backgroundColor: typeSecondaryBackgroundColorHover,
-              },
-            }),
-          }),
+      ...(disabled && {
+        ...disabledStyle,
+        ':hover': {
+          borderColor: undefined,
+          backgroundColor: undefined,
+        },
+      }),
     }
   },
+
+  content: ({ props }) => ({
+    overflow: 'hidden',
+    ...(typeof props.content === 'string' && truncateStyle),
+  }),
 }
 
 export default buttonStyles